refactor: root '/' css files
