+++ src/app/_globals.css
