+++ src/app/page.jsx
@@ -1,5 +1,5 @@
 import Header from '@/components/Header'
-import styles from './main.module.css'
+import styles from './_root.module.css'
 
 export default function Home() {
   return (
